+++ NEWS.md
@@ -1,4 +1,5 @@
-# whatr 0.0.1
+# whatr 0.0.2
 
-* Added basic functionality for scores and board.
+* Added basic functionality for scores.
 * Added a `NEWS.md` file to track changes to the package.
+* Removed all other functions, need to code as `dplyr` pipes.